Start working on better vectors and quaternions
The primary goal is to make these base types more flexible and cross platform. In order to do that, I need to make their basis on structs instead of classes to save memory and boost performance. I have also defined better operators and have tried to make them be more sophisticated in terms of efficiency